pyproject.toml
[project]
name = "tabstash"
version = "0.1.0"
description = "A static guitar tab site generator with mobile-first design"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-frontmatter>=1.1.0",
    "pydantic>=2.5",
    "jinja2>=3.1",
    "markdown>=3.5",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.3",
]

[project.scripts]
tabstash = "tabstash.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tabstash"]

[tool.hatch.build.targets.sdist]
include = ["src/tabstash"]

[tool.hatch.build]
sources = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["--import-mode=importlib"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
